+++ src/main/resources/changelog.md
@@ -1,3 +1,8 @@
+## 1.3.0
+- Added detection for banking elite clues from the ToA reward chest
+- Added "Show next increase" tooltip to clue caskets
+- Added tier-specific clue reward unlocks to the tooltip
+
 ## 1.2.1
 - Fixed issue with infoboxes incorrectly showing full clue stacks on login.